+++ api/src/controllers/iframe.ts
@@ -229,48 +229,44 @@ router.get("/widget/:widgetId/msearch", async (req: Request, res: Response, next
     // If location is set in widget, show only missions in this location
     if (widget.location && widget.location.lat && widget.location.lon) {
       const distance = getDistanceKm(widget.distance && widget.distance !== "Aucun" ? widget.distance : "50km");
-      where.$or = [
-        { geoPoint: { $nearSphere: { $geometry: { type: "Point", coordinates: [widget.location.lon, widget.location.lat] }, $maxDistance: distance * 1000 } } },
-        { "addresses.geoPoint": { $nearSphere: { $geometry: { type: "Point", coordinates: [widget.location.lon, widget.location.lat] }, $maxDistance: distance * 1000 } } },
-      ];
-      whereAggs.$or = [
-        { geoPoint: { $geoWithin: { $centerSphere: [[widget.location!.lon, widget.location!.lat], distance / EARTH_RADIUS] } } },
-        { "addresses.geoPoint": { $geoWithin: { $centerSphere: [[widget.location!.lon, widget.location!.lat], distance / EARTH_RADIUS] } } },
-      ];
+      where["addresses.geoPoint"] = {
+        $nearSphere: {
+          $geometry: { type: "Point", coordinates: [widget.location.lon, widget.location.lat] },
+          $maxDistance: distance * 1000,
+        },
+      };
+      whereAggs["addresses.geoPoint"] = {
+        $geoWithin: {
+          $centerSphere: [[widget.location!.lon, widget.location!.lat], distance / EARTH_RADIUS],
+        },
+      };
     } else if (query.data.lat && query.data.lon) {
       const distance = getDistanceKm("50km");
       // For the aggs, we don't need to check if the remote is set or no, we want all the remote facets
       whereAggs.$and.push({
-        $or: [
-          { geoPoint: { $geoWithin: { $centerSphere: [[query.data.lon, query.data.lat], distance / EARTH_RADIUS] } } },
-          { "addresses.geoPoint": { $geoWithin: { $centerSphere: [[query.data.lon, query.data.lat], distance / EARTH_RADIUS] } } },
-          { remote: "full" },
-        ],
+        $or: [{ "addresses.geoPoint": { $geoWithin: { $centerSphere: [[query.data.lon, query.data.lat], distance / EARTH_RADIUS] } } }, { remote: "full" }],
       });
       // If remote is set to no, show only missions in this location
       if (query.data.remote && query.data.remote.includes("no") && !query.data.remote.includes("yes")) {
-        where.$or = [
-          { geoPoint: { $nearSphere: { $geometry: { type: "Point", coordinates: [query.data.lon, query.data.lat] }, $maxDistance: distance * 1000 } } },
-          { "addresses.geoPoint": { $nearSphere: { $geometry: { type: "Point", coordinates: [query.data.lon, query.data.lat] }, $maxDistance: distance * 1000 } } },
-        ];
+        where["addresses.geoPoint"] = {
+          $nearSphere: {
+            $geometry: { type: "Point", coordinates: [query.data.lon, query.data.lat] },
+            $maxDistance: distance * 1000,
+          },
+        };
       } else if (query.data.remote && query.data.remote.includes("yes") && !query.data.remote.includes("no")) {
         where.remote = "full";
         // Else show missions in this location and remote once
       } else {
         where.$and.push({
-          $or: [
-            { geoPoint: { $geoWithin: { $centerSphere: [[query.data.lon, query.data.lat], distance / EARTH_RADIUS] } } },
-            { "addresses.geoPoint": { $geoWithin: { $centerSphere: [[query.data.lon, query.data.lat], distance / EARTH_RADIUS] } } },
-            { remote: "full" },
-          ],
+          $or: [{ "addresses.geoPoint": { $geoWithin: { $centerSphere: [[query.data.lon, query.data.lat], distance / EARTH_RADIUS] } } }, { remote: "full" }],
         });
       }
     }
 
     // When converting geoPoint for count
-    if (where.geoPoint || where["addresses.geoPoint"]) {
-      const geoWithin = whereAggs.$or ? whereAggs.$or : [{ geoPoint: whereAggs.geoPoint }, { "addresses.geoPoint": whereAggs["addresses.geoPoint"] }];
-      delete where.geoPoint;
+    if (where["addresses.geoPoint"]) {
+      const geoWithin = whereAggs.$or ? whereAggs.$or : [{ "addresses.geoPoint": whereAggs["addresses.geoPoint"] }];
       delete where["addresses.geoPoint"];
       where.$or = geoWithin;
     }
@@ -282,7 +278,6 @@ router.get("/widget/:widgetId/msearch", async (req: Request, res: Response, next
 
     const missions = await MissionModel.find(where).sort({ remote: -1 }).limit(query.data.size).skip(query.data.from).lean();
 
-    if (where.geoPoint) where.geoPoint = whereAggs.geoPoint; // $nearSphere is not supported in countDocuments
     const total = await MissionModel.countDocuments(where);
     const facets = await MissionModel.aggregate([{ $match: whereAggs }, { $facet }]);
