+++ widget-benevolat/pages/index.js
@@ -10,7 +10,7 @@ import { API_URL, DOMAINES, ENV } from "../config";
 import { Carousel } from "../components/carousel";
 import { Grid } from "../components/grid";
 import { Filters, MobileFilters } from "../components/filters";
-
+import { calculateDistance } from "../utils";
 /**
  * Layout widget --> max-width: 1152px
  * 1 : CAROUSEL
@@ -241,16 +241,5 @@ export const getServerSideProps = async (context) => {
   }
   return { props: { widget, missions: [], total: 0, options: {} } };
 };
-
-const calculateDistance = (lat1, lon1, lat2, lon2) => {
-  const R = 6371;
-  const dLat = toRad(lat2 - lat1);
-  const dLon = toRad(lon2 - lon1);
-  const a = Math.sin(dLat / 2) * Math.sin(dLat / 2) + Math.cos(toRad(lat1)) * Math.cos(toRad(lat2)) * Math.sin(dLon / 2) * Math.sin(dLon / 2);
-  const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
-  return R * c;
-};
-
-const toRad = (value) => (value * Math.PI) / 180;
 
 export default Home;
